Refactor App component and move Footer component
