+++ src/App.tsx
@@ -1,16 +1,17 @@
 import { Outlet } from 'react-router-dom';
 import './App.css';
-import { Footer } from './Pages/Footer';
+import { Footer } from './components/Footer';
+import Header from './components/Header';
 
 function App() {
   return (
-    <main>
-      {/* HEADER GOES HERE */}
-
-      <Outlet />
-
+    <>
+      <Header />
+      <main>
+        <Outlet />
+      </main>
       <Footer />
-    </main>
+    </>
   );
 }
